# docs/swagger.yaml
basePath: /
definitions:
  main.Adult:
    properties:
      birth_year:
        example: 1990
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://example.com/photo.jpg
        type: string
      last_name:
        example: Doe
        type: string
      name:
        example: John
        type: string
    type: object
  main.Child:
    properties:
      birth_year:
        example: 2015
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      image_url:
        example: https://example.com/child_photo.jpg
        type: string
      last_name:
        example: Doe
        type: string
      name:
        example: Jane
        type: string
    type: object
  main.Response:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API for managing Adults and Children
  title: ApiGlobal
  version: "1.0"
paths:
  /Add/Adults:
    post:
      consumes:
      - application/json
      description: Add a new adult
      parameters:
      - description: Adult Object
        in: body
        name: adult
        required: true
        schema:
          $ref: '#/definitions/main.Adult'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
      summary: Add a new adult
      tags:
      - Adults
  /Add/Children:
    post:
      consumes:
      - application/json
      description: Add a new child
      parameters:
      - description: Child Object
        in: body
        name: child
        required: true
        schema:
          $ref: '#/definitions/main.Child'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Response'
      summary: Add a new child
      tags:
      - Children
  /Adults:
    get:
      description: Get a list of adults
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Adult'
            type: array
      summary: Get all adults
      tags:
      - Adults
  /Adults/{id}:
    get:
      description: Get a specific adult by ID
      parameters:
      - description: Adult ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Adult'
      summary: Get adult by ID
      tags:
      - Adults
  /Children:
    get:
      description: Get a list of children
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Child'
            type: array
      summary: Get all children
      tags:
      - Children
  /Children/{id}:
    get:
      description: Get a specific child by ID
      parameters:
      - description: Child ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Child'
      summary: Get child by ID
      tags:
      - Children
swagger: "2.0"